Stop relying on environment variables LPC_DEV and LPC_PROJ.
The paths they describe are all the same, relative to the Makefile location.  So
set those paths in the Makefile instead of making the user set new environment
variables.  If the user wants to move directories around, they can override the
variables on the command line, like so:

make LPC_DEV=/path/to/lpc-kit LPC_PROJ=/path/to/lpc-kit/Dev/2148

While we're fiddling with the prelude, remove whitespace at the beginning of the
file and change the description to something a new member of PSAS can
understand.  Not everyone will grok that bob4 is the video overlay board.

Signed-off-by: Sarah Sharp <saharabeara@gmail.com>